tools/requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
openpyxl>=3.1.0
python-telegram-bot>=21.0
